Fix RLE decode bugs
Hard to trace RLE decoding bug was randomly creaing images with bad crops and wrong offsets.
In addition to making the correction. Removed hard to understand and extraneous casting of ints.
Since the byte data is always default read in base 10, it's not neccessary to cast and makes it clearer/natural for contribs and catching future bugs.